+++ R/get_psi_samples.R
@@ -1,9 +1,12 @@
-#' Get samples
-#'
 #' Get samples from PSI input table
 #'
+#' @details
+#'
+#' When \code{config} is provided, then the samples from the \emph{SampleName}
+#' column are used to find the indices in the PSI table.
+#'
 #' @param df A data frame of PSI input values.
-#' @param configdf A 4-column data frame.
+#' @param config Config data frame
 #' @param value If \code{TRUE}, then returns the sample names. Otherwise, return
 #' the indices.
 #' @return If \code{value = TRUE}, A character vector consisting of sample names.
@@ -13,15 +16,16 @@
 #' \dontrun{
 #' get_psi_samples(psi)
 #' }
-get_psi_samples <- function(df, configdf = NULL, value = TRUE) {
-  ix <- seq(7, ncol(df), 2)
-  samples <- colnames(df)[ix]
-  if (!is.null(configdf)) {
-    if (!is.data.frame(configdf)) {
+get_psi_samples <- function(df, config = NULL, value = TRUE) {
+  sample_cols <- colnames(df[, 7:ncol(df)])
+  samples <- grep("(\\.|-)Q$", sample_cols, invert = TRUE, value = TRUE)
+  # samples <- colnames(df)[ix]
+  if (!is.null(config)) {
+    if (!is.data.frame(config)) {
       stop("configdf is not a data frame!")
     }
 
-    samples <- configdf$SampleName[which(configdf$SampleName %in% samples)]
+    samples <- config$SampleName[which(config$SampleName %in% samples)]
   }
   if (value) {
     return(samples)